% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f8_boletin.R
\name{f8_boletin}
\alias{f8_boletin}
\title{Macro 1}
\usage{
f8_boletin(directorio, mes, anio, tipo = "pdf")
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada}

\item{mes}{Definir el mes a ejecutar, ej: 11}

\item{anio}{Definir el año a ejecutar, ej: 2022}
}
\description{
Con esta función realizaremos la primera parte del proceso, en donde se correran las 4
primeras funciones.
}
\details{
Esta funcion tiene como objetivo correr desde la función inicial hasta la función
de identificacion de alertas, con el fin de que el usuario pueda obtener rapidamente el
archivo de salida de la función de alertas y poder realizar el proceso de critica.
Recuerde que si el archivo de alertas ya está creado, al correr esta función le pedira que
responda S si quiere sobreescribirlo, de lo contrario otro valor para cancelar la función

Ver:\code{\link{f0_inicial}},\code{\link{f1_integracion}}, \code{\link{f2_estandarizacion}},
\code{\link{f3_identificacion_outliers}}
}
\examples{
macro1(directorio="Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes=11,anio=2022)
}
