% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f9_shiny.R
\name{f9_shiny}
\alias{f9_shiny}
\title{Shiny alertas}
\usage{
f9_shiny(directorio, mes, anio)
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada}

\item{mes}{Definir el mes a ejecutar, ej: 11}

\item{anio}{Definir el año a ejecutar, ej: 2022}
}
\description{
Está función abre el navegador para abrir la aplicación de alertas, en la cual se puede loguear
y se encontraran 4 ventanas

1. identificación de alertas: la cual se puede filtrar por ID_NUMORD, mes y año; se encontrara una tabla con
los valores para las variables de capitulo 2 y 3 para los filtros aplicados, grafico de la serie para la variable
que se seleccione y finalmente los datos para el establecimiento en esa variable desde 2018 hasta la fecha.

2. Consolidado de alertas: la cual se puede filtrar por capitulo a la que pertenece la variable y el tipo de
imputación (imputación deuda o imputación caso especial) en el año y mes de interes, tiene el resgistro de todos
los establecimientos y variables en los que se identifico un posible caso de imputación.

3. Agregado variables: la cual se puede filtrar por año, mes, actividad economicama, departamento, área metropolitana
y ciudad...

4. Comparación: la cual se puede filtrar por comparación (CIUU,DPTO,ciudad y área metropolitana) , año, mes, actividad economica
departamento, área metropolitana y ciudad...
}
\examples{
f9_shiny(directorio"Documents/DANE/Procesos DIMPE /PilotoEMMET",mes = 11  ,anio =2022)
}
