% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alertas.R
\name{identificacion_outliers}
\alias{identificacion_outliers}
\title{Title}
\usage{
identificacion_outliers(directorio, year, mes)
}
\description{
@param mes Definir las tres primeras letras del mes a ejecutar, ej: Nov
 @param anio Definir el año a ejecutar, ej: 2022
 @param directorio definir el directorio donde se encuentran ubicado los datos de entrada
}
\details{
@return CSV file
 @export

 @examples identificacion_outliers(directorio="/Users/nataliaarteaga/Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes="nov",anio=2022)
}
