% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integracion.R
\name{integracion}
\alias{integracion}
\title{Integración}
\usage{
integracion(directorio, mes, anio)
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada}

\item{mes}{Definir el mes a ejecutar, ej: 11}

\item{anio}{Definir el año a ejecutar, ej: 2022}
}
\value{
CSV file
}
\description{
Con esta función se unificarán todas las bases de datos que son necesarias en el proceso, para
 al final exportar en una única base
}
\examples{
integracion(directorio="/Users/nataliaarteaga/Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes=11,anio=2022)
}
