% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anexonal.R
\name{anacional}
\alias{anacional}
\title{Title}
\usage{
anacional(
  directorio = "/Users/nataliaarteaga/Documents/DANE/Procesos DIMPE /PilotoEMMET",
  mes = "noviembre",
  year = 2022
)
}
\description{
@param mes Definir las tres primeras letras del mes a ejecutar, ej: Nov
 @param anio Definir el año a ejecutar, ej: 2022
 @param directorio definir el directorio donde se encuentran ubicado los datos de entrada
}
\details{
@return CSV file
 @export

 @examples identificacion_outliers(directorio="/Users/nataliaarteaga/Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes="nov",anio=2022)
}
