% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2_estandarizacion.R
\name{estandarizacion}
\alias{estandarizacion}
\title{Estandarizacion}
\usage{
estandarizacion(directorio, mes = 11, anio = 2022)
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada}

\item{mes}{Definir el mes a ejecutar, ej: 11}

\item{anio}{Definir el año a ejecutar, ej: 2022}
}
\value{
CSV file
}
\description{
Esta funcion utiliza como insumo la base de datos integrada, es decir la
 base que denominamos: "Base Panel". Tiene  como objetivo analizar el periodo de tiempo
 en el que cada empresas reporta la producción, en la encuesta, y con ello estandarizar
 las  fechas vacías, tanto de inicio y fin, para que se mantenga el mismo margen de reporte.
 Adicionalmente, en las variables numeticas, los datos faltantes se cambian por cero.
 A estas mismas variables se es establece un formato numerico para la porterior manipulación.
 Finalmente se exporta un archivo de tipo .csv con la información actualizada de las fechas.
}
\details{
Esta funcion recorre las siguiente variables de fecha de inicio y fin:

 II_PA_PP_IEP, II_PA_PP_FEP.
 II_PA_TD_IET, II_PA_TD_FET.
 II_PA_TI_IETA, II_PA_TI_FETA.
 II_PA_AP_AI_AP, II_PA_AP_AF_AP.
 II_PP_PP_IOP, II_PP_PP_FOP.
 II_PP_TD_IOT, II_PP_TD_FOT.
 II_PP_TI_IOTA, II_PP_TI_FOTA.
 II_PP_AP_AI_PP, II_PP_AP_AF_PP.
 II_HORAS_HORDI_D, II_HORAS_HORDI_H.
 III_PE_IV, III_PE_FV.



 La finalidad de la funcion es estandarizar el periodo de las fechas
 en que las empresas generan el reporte mensual.

 Para ello, se analiza el reporte del mes anretior y se calculan el numero de
 días de este. Con ese dato se reemplazan las fehas faltantes del mes presente,
 o las reportados como cero.

 Con esto se logra una unificación del periodo de repote de las empresas.
 En caso dado de no tener reporte del mes anterior se asigna un reporte mensual.
}
\examples{
integracion(directorio="Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes=11,anio=2022)

}
