% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estandarizacion.R
\name{estandarizacion}
\alias{estandarizacion}
\title{Estandarización}
\usage{
estandarizacion(directorio, mes, anio)
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada}

\item{mes}{Definir  mes a ejecutar, ej: 11}

\item{anio}{Definir el año a ejecutar, ej: 2022}
}
\value{
CSV file
}
\description{
Analizar el periodo de tiempo reportado por cada establecimiento, para estandarizar las  fechas vacías, tanto de inicio y fin, para que se mantenga el mismo margen de reporte.
}
\examples{
estandarizacion(directorio="Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes=11,anio=2022)
}
