% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f8_boletin.R
\name{f8_boletin}
\alias{f8_boletin}
\title{Boletin}
\usage{
f8_boletin(directorio, mes, anio, tipo = "pdf")
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada}

\item{mes}{Definir el mes a ejecutar, ej: 11}

\item{anio}{Definir el año a ejecutar, ej: 2022}

\item{tipo}{Definir si se quiere un archivo pdf, word, o html, por defecto esta la opción "pdf", ej: "html}
}
\description{
Está función crea el boletín, un documento exportado en pdf o word, según se desea.

 Este es un resumen ejecutivo en el que se presentan los
 principales resultados con el uso de cuadros de líneas para presentar
 información histórica y graficas de barras para hacer comparaciones de
 variaciones entre las distintas actividades o desagregación regional.
 Además, se presentan tablas con el resumen por actividad para el total
 nacional, departamentos, áreas metropolitanas y ciudades incluyendo
 las tres variables principales que permiten hacer un análisis cruzado
 entre estas.
}
\examples{
f8_boletin(directorio"Documents/DANE/Procesos DIMPE /PilotoEMMET",mes = 11  ,anio =2022,tipo="word")
}
