% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alertas.R
\name{identificacion_outliers}
\alias{identificacion_outliers}
\title{Identificación de alertas}
\usage{
identificacion_outliers(directorio, mes, anio)
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada.}

\item{mes}{Definir el mes a ejecutar, ej: 11}

\item{anio}{Definir el año a ejecutar, ej: 2022}
}
\value{
CSV file
}
\description{
Función para identificar que variables dentro de cada establecimiento  pueden llegar a ser casos
de imputación, ya sean imputación deuda o por casos especiales.

\itemize{

\item Imputación deuda: son todos los establecimientos cuya novedad en la base original tienen valor de 5
(deuda), luego después de la imputación el código de novedad pasa a ser 98 (imputado por deuda).

\item Casos especiales: son aquellos establecimientos de las cuales se detectó que el dato informado
no es lo esperado, se llama al establecimiento para confirmar que el dato reportado es el que es, en dado caso que no exista respuesta por parte de ellos se toma el dato como un caso especial y se procede a imputar.
}
}
\details{
Primero se dividen las variables en capitulo 2 y capitulo 3

\itemize{

\item Variables capítulo 2

-II_PA_PP__NPERS_EP        |   -AJU_II_PA_PP_SUELD_EP

-II_PA_TD__NPERS_ET        |   -AJU_II_PA_TD_SUELD_ET

-II_PA_TI__NPERS_ETA       |   -AJU_II_PA_TI_SUELD_ETA

-II_PA_AP__AAEP            |   -AJU_II_PA_AP_AAS_AP

-II_PP_PP__NPERS_OP        |   -AJU_II_PP_PP_SUELD_OP

-II_PP_TD__NPERS_OT        |   -AJU_II_PP_TD_SUELD_OT

-II_PP_TI__NPERS_OTA       |   -AJU_II_PP_TI_SUELD_OTA

-II_PP_AP__APEP            |   -AJU_II_PP_AP_AAS_PP

-AJU_II_HORAS_HORDI_T      |    -AJU_II_HORAS_HEXTR_T

\item Variables capítulo 3

-AJU_III_PE_PRODUCCION

-AJU_III_PE_VENTASIN

-AJU_III_PE_VENTASEX

-III_EX__VEXIS

}
El procedimiento para identificar posibles establecimientos a imputar es:

1.	Si la novedad es igual 5 sabemos que el individuo se va a imputar por deuda en todas sus variables o si el valor reportado en la variable de interés es cero y el mes inmediatamente anterior fue diferente de cero también se identifica como imputación deuda

2.	Si la novedad es diferente a 5 se observa la variación con respecto al mes anterior; si es menor o igual a 20\% no se imputa.

3.	Si la variación es mayor a 20% y su valor es diferente de cero, entonces, primero se observa
si el dato es igual a alguno de los datos reportados por el establecimiento en meses anteriores;
en caso de que esté presente, se decide no imputar, si el dato no estuvo anteriormente, entonces
se procede a identificar si la variable es referente a capítulo 2, si es así,
se realiza una carta de control con los ultimos 24 meses, si esta identifica el valor como atípico
se imputa por caso especial; si la variable es referente a capítulo 3 ,entonces, se realiza una prueba
mezclando el comando \code{\link[tsoutliers]{locate.outliers.iloop}} y la carta de control de 24 meses, el comando locate.outliers.iloop identifica los valores que
pueden ser outliers en series de tiempo, si el valor fue identificado como un valor atípico se imputa
por caso especial, en caso de que no fuera identificado como un valor atípico no se imputa.

Los limites de la carta control son calculados de la siguiente manera:
\deqn{(\bar{x}−1.96∗𝑆,\bar{x}+1.96∗𝑆)}

Donde \eqn{\bar{x}} es el promedio de la serie de los ultimos 24 meses de la variable y establecimiento de interés y \eqn{S} es la desviación estándar de la serie de los ultimos 24 meses de la variable y establecimiento de interés.
}
\examples{
identificacion_outliers(directorio="Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes=11,anio=2022)

}
