% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alertas.R
\name{identificacion_outliers}
\alias{identificacion_outliers}
\title{Identificaci칩n de alertas}
\usage{
identificacion_outliers(directorio, year, mes)
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada.}

\item{mes}{Definir el mes a ejecutar, ej: 11}

\item{anio}{Definir el a침o a ejecutar, ej: 2022}
}
\value{
CSV file
}
\description{
Funci칩n para identificar establecimientos que pueden llegar a ser casos de imputaci칩n.
}
\examples{
identificacion_outliers(directorio="/Users/nataliaarteaga/Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes=11,anio=2022)

}
