% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputacion.R
\name{imputacion_outliers}
\alias{imputacion_outliers}
\title{Imputaci칩n outliers}
\usage{
imputacion_outliers(directorio, year, mes)
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada}

\item{mes}{Definir las tres primeras letras del mes a ejecutar, ej: 11}

\item{anio}{Definir el a침o a ejecutar, ej: 2022}
}
\value{
CSV file
}
\description{
Funci칩n para realizar la imputaci칩n por deuda y por casos especiales
}
\examples{
imputacion_outliers(directorio="/Users/nataliaarteaga/Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes="nov",anio=2022)
}
